+++ RF_Cascade_wMBS_Predict/RF_Cascade_wMBS_Predict.cxx
@@ -49,14 +49,14 @@ int run_main(int argc, const char **argv)
     double smoothing_scale = 3.0;
     int numGDsteps = atoi(argv[11]);
     MultiArray<1, double> priorStrength(6);
-    priorStrength(1) = atoi(argv[12]);
-    priorStrength(2) = atoi(argv[13]);
-    priorStrength(3) = atoi(argv[14]);
-    priorStrength(4) = atoi(argv[15]);
-    priorStrength(5) = atoi(argv[16]);
-    priorStrength(6) = atoi(argv[17]);
+    priorStrength(1) = atof(argv[12]);
+    priorStrength(2) = atof(argv[13]);
+    priorStrength(3) = atof(argv[14]);
+    priorStrength(4) = atof(argv[15]);
+    priorStrength(5) = atof(argv[16]);
+    priorStrength(6) = atof(argv[17]);
     int numOffsets = atoi(argv[18]);
-    double offsetScale = atoi(argv[19]);
+    double offsetScale = atof(argv[19]);
     float lambdaU = 4;
     float lambdaPW = 4;
